Use singular table names for JDBC authentication
JDBC authentication (that is, authentication from a database), by
default, uses plural table names. This is simply because the SQL queries
in
    org.springframework.security.provisioning.JdbcUserDetailsManager
and in
    org.springframework.security.core.userdetails.jdbc.JdbcDaoImpl
classes uses hard-coded table names, where these tables names are
plural.

To change this situation, we simply need to change the SQL queries in
two aforementioned classes. We do this simply by manually creating a new
instance of 'JdbcUserDetailsManager' and using the setters on this new
instance in order to set the SQL queries.

Note that this new instance of 'JdbcUserDetailsManager' must be defined
as a @Bean and it must be inside a @Configuration. Check out the commit
for more details.